backend/requirements.txt
# Backend Dependencies for DocuSense AI - Version simplifiée

# FastAPI and ASGI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# AI Providers
openai>=1.3.0
anthropic>=0.7.0
mistralai>=0.0.10
ollama>=0.1.0

# File Processing
python-docx>=1.1.0
PyPDF2>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0
redis>=5.0.0
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
slowapi>=0.1.0

# Performance
gunicorn>=21.0.0

# Monitoring
prometheus-client>=0.19.0
structlog>=23.0.0 